+++ driver-core/src/main/java/com/datastax/driver/core/SessionManager.java
@@ -15,7 +15,6 @@
  */
 package com.datastax.driver.core;
 
-import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
 import java.util.*;
@@ -29,8 +28,6 @@ import com.google.common.util.concurrent.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-
-import com.datastax.driver.core.exceptions.AuthenticationException;
 import com.datastax.driver.core.exceptions.DriverInternalError;
 import com.datastax.driver.core.exceptions.UnsupportedFeatureException;
 import com.datastax.driver.core.policies.LoadBalancingPolicy;
@@ -504,6 +501,12 @@ class SessionManager extends AbstractSession {
         DefaultResultSetFuture future = new DefaultResultSetFuture(this, msg);
         execute(future, statement);
         return future;
+    }
+
+    void trashIdleConnections(long now) {
+        for (HostConnectionPool pool : pools.values()) {
+            pool.trashIdleConnections(now);
+        }
     }
 
     private static class State implements Session.State {